# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Booking:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/domain.BookingAlternative'
        type: array
      comment:
        type: string
      completed_at:
        type: string
      confirmed_at:
        type: string
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      guests_count:
        type: integer
      id:
        type: string
      rejected_at:
        type: string
      restaurant_id:
        type: string
      status:
        $ref: '#/definitions/domain.BookingStatus'
      time:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.BookingAlternative:
    properties:
      accepted_at:
        type: string
      booking_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      message:
        type: string
      rejected_at:
        type: string
      time:
        type: string
    type: object
  domain.BookingStatus:
    enum:
    - pending
    - confirmed
    - rejected
    - cancelled
    - completed
    type: string
    x-enum-varnames:
    - BookingStatusPending
    - BookingStatusConfirmed
    - BookingStatusRejected
    - BookingStatusCancelled
    - BookingStatusCompleted
...
  handlers.CreateBookingRequest:
    properties:
      comment:
        type: string
      date:
        type: string
      duration:
        minimum: 30
        type: integer
      guests_count:
        minimum: 1
        type: integer
      restaurant_id:
        type: string
      time:
        type: string
      user_id:
        type: string
    required:
    - date
    - duration
    - guests_count
    - restaurant_id
    - time
    - user_id
    type: object
...
  handlers.RejectBookingRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  handlers.SuggestAlternativeTimeRequest:
    properties:
      date:
        type: string
      message:
        type: string
      time:
        type: string
    required:
    - date
    - time
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant-booking.com
    name: API Support
  description: API для системы бронирования ресторанов
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Booking API
  version: "1.0"
paths:
  /bookings:
    post:
      consumes:
      - application/json
      description: Create a new booking for a restaurant
      parameters:
      - description: Booking data
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant or user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Not enough seats at the specified time
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create booking
      tags:
      - bookings
  /bookings/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get booking
      tags:
      - bookings
  /bookings/{id}/alternative:
    post:
      consumes:
      - application/json
      description: Restaurant suggests an alternative time for a booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Alternative time data
        in: body
        name: alternative_time
        required: true
        schema:
          $ref: '#/definitions/handlers.SuggestAlternativeTimeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot suggest alternative time in current status
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Suggest alternative time
      tags:
      - bookings
  /bookings/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a booking by the user
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot cancel booking in current status
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel booking
      tags:
      - bookings
  /bookings/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a booking as completed
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot complete booking in current status
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Complete booking
      tags:
      - bookings
  /bookings/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Confirm a booking by the restaurant
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot confirm booking in current status
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm booking
      tags:
      - bookings
  /bookings/{id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a booking by the restaurant with a reason
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: reason
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot reject booking in current status
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reject booking
      tags:
      - bookings
  /bookings/alternatives/{id}/accept:
    post:
      consumes:
      - application/json
      description: User accepts the suggested alternative time
      parameters:
      - description: Alternative ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alternative not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accept alternative
      tags:
      - bookings
  /bookings/alternatives/{id}/reject:
    post:
      consumes:
      - application/json
      description: User rejects the suggested alternative time
      parameters:
      - description: Alternative ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alternative not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reject alternative
      tags:
      - bookings
...
schemes:
- http
- https
swagger: "2.0"